combine compliance tests
